requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
jinja2==3.1.3
google-generativeai==0.8.3
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
